+++ __tests__/src/app/page/snapshot.jsx
@@ -3,7 +3,5 @@ import Home from '@/app/page';
 
 it('renders homepage unchanged', () => {
   const { container } = render(<Home />);
-  var a;
-
   expect(container).toMatchSnapshot();
 });
