+++ src/components/Input/interfaces.ts
@@ -1,0 +1,4 @@
+export interface IInputProps {
+  value: string;
+  onChange: (value: string) => void;
+}
